fix: first path component in winbar showing empty string
E.g. ` > etc > passwd`

This happens when `path` starts with `/` e.g. "/etc/passwd". On using
`vim.split` function to split path into components using `/` as
separator, the first component will be an empty string. Thus when path
components are concatenated, the path section will be ` > etc > passwd`.

This is resolved by checking if the first splitted component is an empty
string. If it is, remove the first component and prepend a `/` to the
new first component (old second component). Thus the final concatenated
path section will be `/etc > passwd`.