fix: winbar path component showing only file icon
This happens when `path` is an empty string such as when `nvim` is
called without any file arguments or when reading from stdin.

This is resolved by checking if `path` is an empty string. If it is, the
final path section will be `[No Name]` similar to how `lualine.nvim`
treats `path` being an empty string.